requirements.txt
aiohttp>=3.9.0
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
pymupdf>=1.24.0  # PDF text extraction (optional)
feedparser